+++ frontend-angular/apps/eworkbench/src/app/modules/shared/filter-sidebar/filter-sidebar.component.ts
@@ -42,7 +42,6 @@ export class FilterSidebarComponent implements OnInit {
 
   public ngOnInit(): void {
     this.cmsService.get$.pipe(untilDestroyed(this)).subscribe(({ maintenance }) => {
-      console.log(maintenance);
       this.cmsMessageShown = maintenance.visible;
       this.cdr.markForCheck();
     });