+++ frontend-angular/apps/eworkbench/src/app/pages/appointments/components/appointments-page/appointments-page.component.ts
@@ -846,10 +846,11 @@ export class AppointmentsPageComponent implements OnInit {
   }
 
   public openNewModal(): void {
+    const initialState = this.project ? { projects: [this.project] } : null;
     /* istanbul ignore next */
     this.modalRef = this.modalService.open(NewAppointmentModalComponent, {
       closeButton: false,
-      data: { withSidebar: this.showSidebar, initialState: { projects: this.project ? [this.project] : [] } },
+      data: { withSidebar: this.showSidebar, initialState: initialState },
     });
     /* istanbul ignore next */
     this.modalRef.afterClosed$.pipe(untilDestroyed(this), take(1)).subscribe((callback: ModalCallback) => this.onModalClose(callback));