+++ frontend-angular/apps/eworkbench/src/app/pages/contacts/components/contacts-page/contacts-page.component.ts
@@ -719,10 +719,11 @@ export class ContactsPageComponent implements OnInit {
   }
 
   public openNewModal(): void {
+    const initialState = this.project ? { projects: [this.project] } : null;
     /* istanbul ignore next */
     this.modalRef = this.modalService.open(NewContactModalComponent, {
       closeButton: false,
-      data: { withSidebar: this.showSidebar, initialState: { projects: this.project ? [this.project] : [] } },
+      data: { withSidebar: this.showSidebar, initialState: initialState },
     });
     /* istanbul ignore next */
     this.modalRef.afterClosed$.pipe(untilDestroyed(this), take(1)).subscribe((callback: ModalCallback) => this.onModalClose(callback));