+++ frontend-angular/apps/eworkbench/src/app/pages/files/components/files-page/files-page.component.ts
@@ -820,10 +820,11 @@ export class FilesPageComponent implements OnInit {
   }
 
   public openNewModal(): void {
+    const initialState = this.project ? { projects: [this.project] } : null;
     /* istanbul ignore next */
     this.modalRef = this.modalService.open(NewFileModalComponent, {
       closeButton: false,
-      data: { withSidebar: this.showSidebar, initialState: { projects: this.project ? [this.project] : [] } },
+      data: { withSidebar: this.showSidebar, initialState: initialState },
     });
     /* istanbul ignore next */
     this.modalRef.afterClosed$.pipe(untilDestroyed(this), take(1)).subscribe((callback: ModalCallback) => this.onModalClose(callback));