+++ frontend-angular/apps/eworkbench/src/app/pages/pictures/components/pictures-page/pictures-page.component.ts
@@ -716,10 +716,11 @@ export class PicturesPageComponent implements OnInit {
   }
 
   public openNewPictureModal(): void {
+    const initialState = this.project ? { projects: [this.project] } : null;
     /* istanbul ignore next */
     this.modalRef = this.modalService.open(NewPictureModalComponent, {
       closeButton: false,
-      data: { withSidebar: this.showSidebar, initialState: { projects: this.project ? [this.project] : [] } },
+      data: { withSidebar: this.showSidebar, initialState: initialState },
     });
     /* istanbul ignore next */
     this.modalRef.afterClosed$.pipe(untilDestroyed(this), take(1)).subscribe((callback: ModalCallback) => this.onModalClose(callback));