+++ frontend-angular/apps/eworkbench/src/app/pages/task-boards/components/task-boards-page/task-boards-page.component.ts
@@ -690,10 +690,11 @@ export class TaskBoardsPageComponent implements OnInit {
   }
 
   public openNewTaskBoardModal(): void {
+    const initialState = this.project ? { projects: [this.project] } : null;
     /* istanbul ignore next */
     this.modalRef = this.modalService.open(NewTaskBoardModalComponent, {
       closeButton: false,
-      data: { withSidebar: this.showSidebar, initialState: { projects: this.project ? [this.project] : [] } },
+      data: { withSidebar: this.showSidebar, initialState: initialState },
     });
     /* istanbul ignore next */
     this.modalRef.afterClosed$.pipe(untilDestroyed(this), take(1)).subscribe((callback: ModalCallback) => this.onModalClose(callback));